+++ src/views/Start.vue
@@ -94,6 +94,12 @@ import { config } from "../assets/config.js";
               Factors that may have contributed to the episode of DKA occuring
               (preventable factors) are now collected for audit purposes.
             </li>
+            <li>
+              When the patient has pre-existing diabetes ticks are automatically added to the not applicable boxes for ongoing insulin therapy for the mode the patient does not use on flow chart 5.
+            </li>
+            <li>
+              There is now a facility to update limited episode data after submission. This is currently limited to preventable factors and pre-existing diabetes status, but may be expanded in the future to allow retrospective audit data collection.
+            </li>
           </ul>
         </div>
       </div>